Add implementations for the Observer Pattern
This commit adds the top-level implementations for the Observer Pattern
which will be used for communications between the Model, View and
Controller. The implementation consists of two parts; the *Observed*
which is the subject whom is capable of pushing events. And the
*Observer* which is the listener which receive events and act upon them.

Patch also contains all tests required by the feature description. The
tests are contained in the *ObserverTests* and *ObservedTests* classes
in the test section. Worth reiterating is that there only exist tests
for methods which provide default implementations. Also with the tests
two additional classes are included: *TestingConcreteObserver* and
*TestingConcreteObserved*.

These classes are concrete implementations of the *Observer* interface,
and the *Observed* abstract class. They are needed to be able to test
their default method, since they must be instantiated. These testing
concrete classes are only used to test the default implementations; and
therefore all of their method implementations are empty, since they are
not needed.