Add Model superclass
This commit implements the *Model* superclass which will server as the
parent element of any concrete *Model* implementations. The model
defines abstract methods for each interaction which a controller may
need to perform with a model; both for executing events and accessing
information.

The *Model* class also extends the *Observed* class to allow observers
(in the future controllers will use this) to attach themselves to listen
for to events which the model can cause. The *Model* class also provides
helper methods for each of the event types which a model can cause; just
to make it easier to use. The helper methods simply pass along their
provided parameters to the Observed.pushEvent() method, along with a
unique event name for each event type.