Add View superclass
This commit implements the *View* superclass which will serve as the
parent element of any concrete *View* implementations. The view defines
abstract methods for each interaction which is required by a controller.
So both for pushing events and for updating elements contained in the
view.

The *View* class also extends the *Observed* class to allow observers
(in the future controllers) to listen for events. The *View* class also
provides helper methods for pushing each event type in a simple manner.
The helper methods simply pass along their parameters to the
*Observed.pushEvent()* method along with the appropriate event type;
removing the need to include it in the main functional code.