Add BackingSquare class
This commit implements the *BackingSquare* class which will be used by
the model to store state information about each individual square. The
class contains three booleans: revealed, flagged and mine; each with
a set of getters and setters. The class provides no "valid state"
checking and instead leaves it up to the model implementation.

The tests included are simple, first there is a test for each getter
which checks that the default value when a instance is created is
correct (false). Then there is another set of test methods which tests
getters and setters in combination to ensure that the correct value is
always set and returned.