Add implementation and tests for the Square GUI component
This adds the JPanel extended square GUI component which represents a single square in the Minesweeper board. The component contains three states which is set by the features, set to flag, set to mine, set to a numbered value or set to hidden. Other than setting the state, the flag and mine feature adds a respective image icon for display. The set to a numbered value adds the provided value for display and the set to hidden removes any icon and numbers displayed for the square.

This also adds tests to ensure correct state changes between the features, ensure the features adds the value/icon image for display, ensure zero number values are not added for display as they are to represent a blank square.