Add Controller superclass
This commit implements the *Controller* abstract superclass which will
server as the parent element of any concrete *Controller*
implementations. The controller defines abstract methods for handling
all of the possible events that can be pushed by either models or views.

To to this the *Controller* implements the *Observer* interface which
will be used to attach to Models and Views to listen for their events.
This causes the method *handleEvent()* to be triggered when an event is
pushed. The *Controller* implements this method and directs each event
type to the appropriate event specific handler.