Implement board generation functionality in concrete model
This commit partially implements the *GameModel* concrete implementation
of the *Model* class. This adds functionality for generating the board
and randomizing the placements of the mines; along with helper methods
to achieve this. Specifically a method for getting all neighbors of a
square and a implementation of the *Model.isMine()* method.

The generation sequence is split into three parts:

First the storage object is initialized, currently a two-dimensional
list of *BackingSquare* objects is used. This is initialized using a
length parameter provided in the object constructor, which decides the
size of the board. Empty nested lists are then initialized to form the
full storage object.

Second the *BackingSquare* objects are initialized. Each possible
location on the board is iterated through and a new *BackingSquare*
instance is created. Then a random number of generated, and together
with a predefined constant *_mineChance*, whether the square is a mine
is set. This step is repeated for every *BackingSquare*.

Lastly, a "first location" parameter is passed to the generation method.
This location is used to ensure that the first square that is clicked is
not a mine, which would cause the user to loose unfairly. To prevent
this a check is run to ensure that the first clicked square, and its
direct neighbors are never allowed to be mines. This results in a
2x2 - 3x3 area always being opened upon first click giving the user a
fair shot.