Fix naming convention of methods in SquareTests and private variables in Square
The naming convention defined in the Code Conventions specify that each private variable in non unit-tests should be prefixed with an underscore "_" and the unit-tests follow specific format: ”test_ThingThatItTested_SpecificTest()”; which was not followed in previous commit.

The validate method was also moved in the setValue and setMine methods to visual update the component after background colour and border change; which was previously done before the change.