Add functionality and tests to generate and display the board in concrete View
This implements a part of the GameView concrete implementation of the View class. Functionality is added to generate a board containing the Square GUI components and initialization of a window frame to display the board.

The Squares in the board is implemented with mouse-listeners to call the View pushSelectEvent and pushFlagEvent methods with a left and right mouse button click. That is if the board is not locked or the square has already been selected.

Tests have been implemented to ensure the board generates the correct amount of square components and the View push methods gets called correctly.