Implement GameModel.selectSquare method
This commit adds an implementation for the *selectSquare* method in the
*GameModel* class. It takes in a location, which will be a square that
was clicked by the user; and determines what happens. If the clicked
square is flagged or already revealed; it does nothing. If the the
clicked square is a mine; a "mine hit" event is triggered.

If the clicked square is a value square; a reveal square event is
triggered. It then calculates all squares related to the original
location which are connected via a "zero value" square. If the original
square is zero, then all neighbors of the square is included. If any
neighbor is zero, the same check is done for it, to include its
neighbors and check for additional zeros; and so on. This causes all
related zero value squares to be revealed which gives the user
additional visibility; and is what makes the game not impossible.