Fix wildcard imports and small misses in GameView component and tests
This commit removes and replaces wildcard imports to specific component imports in the GameView and GameViewTests components. The _frame and _board Swing variables in GameView is no longer initialized outside the specific initiation methods. The _frame variables pack() method is moved down in the initializeFrame() method and the nested if-statements in the selectSquare() and flagSquare() methods is merged into one.

In the GameViewTests unit-test component the test methods that tests the View push events are expanded to give better test coverage.