Implement GameView Square set methods with tests
This commit implements the GameView set methods for executing a specific square from the boards set methods. To access the specific Square from the board using a point position the getSquareFromPosition() method was created.

Tests were also created to make sure the getSquareFromPosition() method returns null if the point location is invalid and tests for the set methods to make sure the squares set methods is executed as long as the square is not null.