Add GameView menu with buttons
This commit adds a menu to the GameView with a text and the two buttons, restart and quit. The menu text which is initially set to the games title "MineSweeper" can be changed later on to display if the game is over or if the game is won. The buttons, which are initially hidden will be utilized to either, restart the game by calling a reset game event or to quit the game with a system exit call. The board and menu components are positioned placed in a game panel component to easier scale and position.

Tests were also created to make sure the initial menu text and button texts are correct. Further the buttons are tested to make sure they are initially hidden and that the restart button calls the reset event.