Implement game over, reset and win GameView methods
This commit implements the gameOver(), reset() and win() methods from View. The gameOver method calls the setMine() method with the provided point position, changes the menu text to "Game over!", shows the menu buttons and locks the board. The win() method changes the menu text to "You win!", shows the menu buttons and also locks the board. The reset() method calls the setHidden() method on all the squares in the board, changes the menu text back to "Minesweeper", hides the buttons again and unlocks the board.

Tests were also created to make sure the above features handles the squares correctly, sets the correct menu text, hides and shows the menu buttons correctly and locks/unlocks the board lock correctly.