Implement GameController with handle event methods and tests
This commit adds the GameController class which attaches itself as an observer for the View and Model. The class implements the concrete controller handle event methods for handling select square events, flag square events, reset game events, reveal square events, mine hit events and win events to handle communication between the components.

Tests for the handle event methods are also added to ensure the event calls the correct View and/or Model method.