Fix window spam caused by missing JFrame disposal in GameView class
This commit adds a destructor method for the *GameView* class which
allows you to programmatically close the JFrame associated with it. This
results in it being possible to actually "delete" the full *GameView*.

Previously, even if all references to the *GameView* object were
removed, the window would remain, since the JFrame will not auto-exit.
Now instead, the method *GameView.destroy* provides this
functionality.

Lastly, *@AfterEach* methods were added to all test classes which
initialize *GameView* objects; and they now call the destroy method
to actually close the windows and reclaim their resources before moving
on to the next test. As previously, the JFrames of all used
*GameView* instances in a test class would remain open until the
entire test class finished.