Implement main method
This commit implements the main method to instantiate a set of MVC
components. The components are initialized using a predefined board
size; and are called using the *invokeLater* method to ensure.

Creating the instances on using *invokeLater* ensures exactly all GUI
modifying calls are run on the main event thread. And since most calls
will be made on the GUI thread (I.E. any call chain started by a user
input event on the GUI) there will be negligible unnecessary load on the
event thread.