Add the report to the repository
This commit adds the full report into the repository. As the report was
originally written in a separate google docs document, all that has been
done for this commit is:

* Reformat the contents to markdown.
* Copy sections into corresponding report sections.

And as a side-node, while the full report is added here in a single
commit; the report was written jointly. But in the modern day it is more
convenient to work simultaneously on a text in a google doc, rather than
via git when possible.